+++ packages/orq-rc/src/orq_ai_sdk/models/agenttoolinputcrud.py
@@ -13,6 +13,7 @@ from .deletememorydocumenttoolinput import (
     DeleteMemoryDocumentToolInput,
     DeleteMemoryDocumentToolInputTypedDict,
 )
+from .filesystemtoolinput import FileSystemToolInput, FileSystemToolInputTypedDict
 from .functiontoolinput import FunctionToolInput, FunctionToolInputTypedDict
 from .googlesearchtoolinput import GoogleSearchToolInput, GoogleSearchToolInputTypedDict
 from .httptoolinput import HTTPToolInput, HTTPToolInputTypedDict
@@ -55,19 +56,20 @@ AgentToolInputCRUDTypedDict = TypeAliasType(
     "AgentToolInputCRUDTypedDict",
     Union[
         CurrentDateToolInputTypedDict,
-        WebScraperToolInputTypedDict,
+        AdvisorToolInputTypedDict,
         CallSubAgentToolInputTypedDict,
         RetrieveAgentsToolInputTypedDict,
         QueryMemoryStoreToolInputTypedDict,
         WriteMemoryStoreToolInputTypedDict,
         RetrieveMemoryStoresToolInputTypedDict,
         DeleteMemoryDocumentToolInputTypedDict,
-        QueryKnowledgeBaseToolInputTypedDict,
         RetrieveKnowledgeBasesToolInputTypedDict,
+        QueryKnowledgeBaseToolInputTypedDict,
+        WebScraperToolInputTypedDict,
         SidekickToolInputTypedDict,
-        AdvisorToolInputTypedDict,
         GoogleSearchToolInputTypedDict,
         CodeInterpreterToolInputTypedDict,
+        FileSystemToolInputTypedDict,
         FunctionToolInputTypedDict,
         JSONSchemaToolInputTypedDict,
         HTTPToolInputTypedDict,
@@ -95,6 +97,7 @@ AgentToolInputCRUD = Annotated[
         Annotated[AdvisorToolInput, Tag("advisor")],
         Annotated[SidekickToolInput, Tag("sidekick")],
         Annotated[CodeInterpreterToolInput, Tag("code_interpreter")],
+        Annotated[FileSystemToolInput, Tag("file_system")],
         Annotated[HTTPToolInput, Tag("http")],
         Annotated[CodeToolInput, Tag("code")],
         Annotated[FunctionToolInput, Tag("function")],